+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Changes since v0.1.1
 
 * Add error message boxes
-* Improve performance of editing large image
+* Improve performance when editing a large image
 * Improve flood fill performance
 * Fix visual glitches during window creation
 * Fix crash when loading palette-based PNG files (automatically converted to 32-bit for now)
@@ -11,3 +11,4 @@
 * Add about dialog
 * Active selection is now visible in the preview widget
 * Compress undo/redo-steps for much lower memory usage
+* Add distinct blend modes replace and alpha blend